完成FindRoad部分的编写

+++ FindRoadDisplay/graph.h
@@ -8,7 +8,7 @@ class FINDROADDISPLAYSHARED_EXPORT Graph
 {
 public:
     int vNum = 0;
-    int edges[MAX_VEX_NUM][MAX_VEX_NUM];
+    float edges[MAX_VEX_NUM][MAX_VEX_NUM];
     int way[MAX_VEX_NUM];//储存最短路径各点下标
     Graph();
     bool getWay(int start, int end);
